+++ web/src/pages/[crate_names].tsx
@@ -5,58 +5,23 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import React, { useEffect, useState } from 'react';
-import { useRecoilValue } from 'recoil';
-import { crateDataResultsState, crateDownloadDataResultsState } from 'recoil/atoms';
-import Head from 'next/head';
-import { useRouter } from 'next/router';
+import React from 'react';
 import { css } from '@emotion/react';
-import InputForm from 'components/shared/InputForm';
-import ExtraInfo from 'components/shared/ExtraInfo';
-import CratesTable from 'components/[crate_names]/CratesTable';
-import DownloadChart from 'components/[crate_names]/DownloadChart';
+import CratesView from 'components/shared/CratesView';
 
 const CratesCompare = (): JSX.Element => {
-  const router = useRouter();
-  const { crate_names } = router.query;
-
-  const [crateNames, setCrateNames] = useState([]);
-  const crateDataResults = useRecoilValue(crateDataResultsState(crateNames));
-  const crateDownloadDataResults = useRecoilValue(crateDownloadDataResultsState(crateNames));
-
-  useEffect(() => {
-    if (!crate_names) return;
-    const crateNameList = Array.from(new Set(crate_names.toString().split('+'))); // unique!
-    setCrateNames(crateNameList);
-  }, [crate_names, setCrateNames]);
-
-  // TODO: it doesnt work after switching to react-query
-  if (crateDataResults.length === 0 || crateDownloadDataResults.length === 0) {
-    return (
-      <div css={PageIndicator}>
-        <div className="loader" />
-      </div>
-    );
-  }
-
   return (
-    <React.Suspense fallback={<div>loading</div>}>
-      <div css={Wrapper}>
-        <Head>
-          <title>{String(crate_names).split('+').join(', ')} | Crate Trends</title>
-        </Head>
-        <InputForm />
-        <DownloadChart downloadsData={crateDownloadDataResults} />
-        <CratesTable cratesData={crateDataResults} />
-        <ExtraInfo />
-      </div>
+    <React.Suspense
+      fallback={
+        <div css={PageIndicator}>
+          <div className="loader" />
+        </div>
+      }
+    >
+      <CratesView />
     </React.Suspense>
   );
 };
-
-const Wrapper = css`
-  flex: 1;
-`;
 
 const PageIndicator = css`
   display: flex;
